ré-organise les fichiers associés aux pages "détail" des indicateurs
