+++ app.territoiresentransitions.react/src/app/pages/collectivite/Indicateurs/IndicateurPersonnalise.tsx
@@ -9,7 +9,7 @@ import {useIndicateurACompleter} from './useIndicateurDefinitions';
 import {useUpsertIndicateurPersoDefinition} from './useUpsertIndicateurPersoDefinition';
 import IndicateurChart from './charts/IndicateurChart';
 import {HeaderIndicateur} from './Header';
-import {IndicateurValuesTabs} from './IndicateurValuesTabs';
+import {IndicateurValuesTabs} from './detail/IndicateurValuesTabs';
 import {FichesActionLiees} from './FichesActionLiees';
 import {useCurrentCollectivite} from 'core-logic/hooks/useCurrentCollectivite';
